# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Social Network Admin API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Campus {
  id: number;
  name: string;
  lat: number;
  lng: number;
  defaultZoom: number;
  createdAt?: string;
  updatedAt?: string;
}

export interface SetCampusRequest {
  name: string;
  lat: number;
  lng: number;
  defaultZoom?: number;
}

/**
 * Type of location
 */
export type LocationType = (typeof LocationType)[keyof typeof LocationType];

export const LocationType = {
  building: "building",
  dining_hall: "dining_hall",
  sports_field: "sports_field",
  parking: "parking",
  green: "green",
  other: "other",
} as const;

export interface LatLng {
  lat: number;
  lng: number;
}

export interface Location {
  id: number;
  campusId: number;
  name: string;
  description?: string;
  type: LocationType;
  color: string;
  adminName?: string;
  lat: number;
  lng: number;
  polygon: LatLng[];
  /** Building name detected from OpenStreetMap */
  osmName?: string;
  createdAt?: string;
  updatedAt?: string;
}

export interface CreateLocationRequest {
  name: string;
  description?: string;
  type: LocationType;
  color?: string;
  adminName?: string;
  lat: number;
  lng: number;
  polygon: LatLng[];
  osmName?: string;
}

export type AnnouncementPriority =
  (typeof AnnouncementPriority)[keyof typeof AnnouncementPriority];

export const AnnouncementPriority = {
  normal: "normal",
  important: "important",
  urgent: "urgent",
} as const;

export interface Announcement {
  id: number;
  locationId: number;
  title: string;
  content: string;
  priority: AnnouncementPriority;
  createdAt: string;
}

export type CreateAnnouncementRequestPriority =
  (typeof CreateAnnouncementRequestPriority)[keyof typeof CreateAnnouncementRequestPriority];

export const CreateAnnouncementRequestPriority = {
  normal: "normal",
  important: "important",
  urgent: "urgent",
} as const;

export interface CreateAnnouncementRequest {
  title: string;
  content: string;
  priority?: CreateAnnouncementRequestPriority;
}

export type DayOfWeek = (typeof DayOfWeek)[keyof typeof DayOfWeek];

export const DayOfWeek = {
  sunday: "sunday",
  monday: "monday",
  tuesday: "tuesday",
  wednesday: "wednesday",
  thursday: "thursday",
  friday: "friday",
  saturday: "saturday",
} as const;

export interface ScheduleEntry {
  id: number;
  locationId: number;
  dayOfWeek: DayOfWeek;
  /** HH:MM format */
  startTime: string;
  /** HH:MM format */
  endTime: string;
  /** Course/event name */
  label: string;
  instructor?: string;
  createdAt?: string;
}

export interface CreateScheduleEntryRequest {
  dayOfWeek: DayOfWeek;
  startTime: string;
  endTime: string;
  label: string;
  instructor?: string;
}

export type DailyMenuItemsItemCategory =
  (typeof DailyMenuItemsItemCategory)[keyof typeof DailyMenuItemsItemCategory];

export const DailyMenuItemsItemCategory = {
  starter: "starter",
  main: "main",
  side: "side",
  dessert: "dessert",
  drink: "drink",
} as const;

export type DailyMenuItemsItem = {
  name: string;
  category: DailyMenuItemsItemCategory;
};

export interface DailyMenu {
  id: number;
  locationId: number;
  date: string;
  items: DailyMenuItemsItem[];
  averageRating: number;
  ratingCount: number;
  createdAt?: string;
}

export type CreateMenuRequestItemsItemCategory =
  (typeof CreateMenuRequestItemsItemCategory)[keyof typeof CreateMenuRequestItemsItemCategory];

export const CreateMenuRequestItemsItemCategory = {
  starter: "starter",
  main: "main",
  side: "side",
  dessert: "dessert",
  drink: "drink",
} as const;

export type CreateMenuRequestItemsItem = {
  name: string;
  category: CreateMenuRequestItemsItemCategory;
};

export interface CreateMenuRequest {
  date: string;
  items: CreateMenuRequestItemsItem[];
}

export interface RateMenuRequest {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  raterName?: string;
}

export type GameSessionSport =
  (typeof GameSessionSport)[keyof typeof GameSessionSport];

export const GameSessionSport = {
  football: "football",
  basketball: "basketball",
  volleyball: "volleyball",
  tennis: "tennis",
  other: "other",
} as const;

export type GameSessionVotesItem = {
  playerName: string;
  votedAt: string;
};

export interface GameSession {
  id: number;
  locationId: number;
  sport: GameSessionSport;
  scheduledAt: string;
  description?: string;
  maxPlayers?: number;
  votes: GameSessionVotesItem[];
  createdAt?: string;
}

export type CreateGameRequestSport =
  (typeof CreateGameRequestSport)[keyof typeof CreateGameRequestSport];

export const CreateGameRequestSport = {
  football: "football",
  basketball: "basketball",
  volleyball: "volleyball",
  tennis: "tennis",
  other: "other",
} as const;

export interface CreateGameRequest {
  sport: CreateGameRequestSport;
  scheduledAt: string;
  description?: string;
  maxPlayers?: number;
}

export interface VoteGameRequest {
  playerName: string;
}