# lib/api-zod/src/generated/types/rateMenuRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Social Network Admin API
 * OpenAPI spec version: 0.2.0
 */

export interface RateMenuRequest {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  raterName?: string;
}